admin must be logged in for user delete to work

+++ cms/admin/includes/user_functions.php
@@ -119,13 +119,15 @@
      function delete_user() {
         global $pdo;
         if (isset($_GET['delete'])) {
-            $user_id = $_GET['delete'];
+            if (isset($_SESSION['user_role']) && $_SESSION['user_role'] === "admin") {
+                $user_id = $_GET['delete'];
 
-            $query = "DELETE FROM users WHERE user_id = :uid";
-            $stmt = $pdo->prepare($query);
-            $stmt->execute(array(':uid' => $user_id));
-            //refresh after delete
-            header("Location: users.php");
+                $query = "DELETE FROM users WHERE user_id = :uid";
+                $stmt = $pdo->prepare($query);
+                $stmt->execute(array(':uid' => $user_id));
+                //refresh after delete
+                header("Location: users.php");
+            }
         }
     }
